Refactor writing of tempate files
All files written by scripts/configure_container_agent.sh to any dir
in /var/run and /var/lib are now written there when the container
starts and no during docker build. Some templates written in
/usr/libexec are moved to scripts/write-os-apply-config-templates are
written when the container image is built. The reason for that is that
since the agent writes in /var/lib and /var/run we mount as writable
these dirs and they are overwritten when the container starts.

Additionally, we mount /etc/ because os-refresh config needs to write
in it.